requirements.txt
connexion[swagger-ui] >= 2.6.0; python_version>="3.6"
swagger-ui-bundle >= 0.0.2
python_dateutil >= 2.6.0
setuptools >= 21.0.0
Flask == 2.1.1
openstacksdk
os_client_config
gunicorn
pymemcache
